ConformU compliance:
- All 10 ASCOM device types pass ConformU 0E/0I
- DeviceState for all devices with latest interface versions
- Telescope: meridian flip, sidereal tracking, SideOfPier convention
- Camera: time-based exposure lifecycle, named gain/offset
- Mock devices track state for conformance testing

CI/Release:
- Switch from release-plz to release-please
- CI matrix with path filtering
- Apache-2.0 license only

Conflicts resolved: all files taken from 030-conformu-testing branch (authoritative).

+++ tests/image_tests.rs
@@ -137,7 +137,7 @@ fn imagebytes_large_image() {
         assert_eq!(result.len(), rows);
         assert_eq!(result[0].len(), cols);
         assert_eq!(result[0][0], 0);
-        assert_eq!(result[767][1023], 767 * 1024 + 1023);
+        assert_eq!(result[767][1023], (767 * 1024 + 1023) as i32);
     } else {
         panic!("Expected I32_2D");
     }
